Updated documentation about transform pattern
Added information about the type of the content parameter. 
The example for transform does not give any clues about the Object type of the `content` parameter. 
Since it's a Buffer object, and a simple transform could be a pattern replacement  using String.match() and String.replace(), it will be useful to know this information for most people, guessing that is a String object and can be manipulated as is.