Improve clarity on how plugin works with built files
It appears that copy-webpack-plugin is designed to copy files _outside_ of the build process, rather than copy files _created by_ the build process. See: https://github.com/webpack-contrib/copy-webpack-plugin/issues/261

If one uses webpack-copy-plugin to try and copy a built file, if the file did not exist already, then the developer may receive a `WARNING in unable to locate` message, and the file is not correctly copied. This can cause some weird behaviour, especially because if you run webpack again, it will work fine (because it's copied over the previously built file).

I would like to suggest adding in this documentation to the README to prevent people from using the plugin when it's not appropriate for their needs.